Fix EOF and client reset errors
The Request method is making two requests - one for unauthed user and
then another with auth if the first fails.  This double request is
leading to EOF errors from the client.  It is also overhead because
each client request ends up making 2 request requests, and
since most our calls are authenticated this is almost alwasy the case.
Note the only unauth request is (Ping) which directly using the http pkg.