K8S-31: Connection Persistence and UUID Checks
Extends the Couchbase client type to have a UUID field and cache the
http.Client so that connections are persisted.  Endpoints are
dynamically set by the CouchbaseClient in the operator to separate HTTP
connection stuff from who to dial in an particular instance.

The main meat of the change (and no one blames you for thinking it's
horrible) is the client creation which overrides Go's default dialing
and TLS handshaking so that when a new connection is dialed we can
optionaly verify the UUID is for the cluster we are managing, and the
node is clustered.

Change-Id: I540f15505220a9c5f37e06b677750cdb5cfbe366
Reviewed-on: http://review.couchbase.org/90524
Reviewed-by: Tommie McAfee <tommie@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>
Tested-by: Simon Murray <simon.murray@couchbase.com>