K8S-53: Alternate Address Support
Most cloud VPNaaS offer site to site L3 connectivity at the node network
level and not spanning pod networks.  This means that XDCR will not be
able to access the necessary Couchbase server ports.  This change
extends the client layer to allow explicit setting of a node IP and
mapped ports which will allow L3 connectivity from the node network.

Change-Id: Ie4f7b13035ad2053b8eaef10f14b14af02d4e9e5
Reviewed-on: http://review.couchbase.org/91695
Reviewed-by: Mike Wiederhold <mike@couchbase.com>
Tested-by: Simon Murray <simon.murray@couchbase.com>