K8S-783: Kill Pods During Rebalance Reliably
We currently have no way of killing pods reliably while they are being
balanced in.  The two strategies are to just terminate after the
rebalance event is raised, or wait a bit after this.  As no vbuckets may
have been transferred to the new nodes, or we wait too long, we end up
with inconsistent behaviour.  By making the rebalance feedback a channel
we can communicate back to the client code what state we are in an more
importantly how far through we are, thus allowing us to kill the pod
exactly when 50% of the vbuckets are moved for example.  Obviously the
numbers reported are never 100%, so you need to use some trial and error
to determine the correct thereshold for your test.

Change-Id: Ib723fdf7ef9ad6d9d6c7dfb71e1a8c462a1f4573
Reviewed-on: http://review.couchbase.org/104081
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>